requirements.txt
Brotli==1.0.9
certifi==2024.6.2
charset-normalizer==2.0.4
click==8.1.7
idna==3.7
pip==24.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
setuptools==69.5.1
urllib3==2.2.2
wheel==0.43.0
